requirements.txt
discord.py>=1.3.4
jishaku>=1.18.2
statcord.py>=2.2.0